Move to /Election directory and added test cases
